requirements.txt
apify>=2.0
curl-cffi>=0.7
beautifulsoup4>=4.12
readability-lxml>=0.8
lxml>=5.0
html2text>=2024.2
pytz
